requirements.txt
gphoto2 >= 1.8.2
schedule >= 0.5.0
PyYAML >= 3.12